章の削除時の確認を追加

+++ src/view/l1/DeleteParagraphButton.tsx
@@ -4,7 +4,11 @@ import styled from 'styled-components';
 import { IconButton, Tooltip } from '@mui/material/';
 import { Delete } from '@mui/icons-material/';
 
+import { RootState } from '../../store/rootReducer';
 import ScenarioSlice, { DeletePayload, } from '../../store/ScenarioSlice';
+import Paragraph from '../../store/model/Paragraph';
+
+import ScenarioUtil from '../../lib/ScenarioUtil';
 
 
 const Main = styled(IconButton)`
@@ -17,12 +21,19 @@ type Props = {
 
 const App = (props: Props) => {
     const dispatch = useDispatch();
+    const paragraph: Paragraph = useSelector((state: RootState) => 
+        state.scenario.paragraphList[props.id]);
 
-    const onClick = () => {
+    const onClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
         const payload: DeletePayload = {
             paragraphId: props.id
         }
-        dispatch(ScenarioSlice.actions.delete(payload));
+        if (ScenarioUtil.isParagraphEmpty(paragraph)
+            || confirm('章を削除します。よろしいですか？')) {
+            dispatch(ScenarioSlice.actions.delete(payload));
+        }
+
+        e.stopPropagation(); // アコーディオンの開閉をしないようにする。
     }
 
     return (
